[utest][ethernet] Fix flaky tests
There were two flaky tests here:
* EthernetLinkStatusTest was relying on a zx_socket signal_peer call to
propagate to the other process and wake up a wait_one call without using
any delays in the test process. In all cases I could see, the signal
does eventually make it, but too late for the test to succeed. Adding a
fixed delay to the test after sending the signal lets the kernel wake
the other process to handle the signal.

* EthernetDataTest_Recv was racing against the devmgr cleaning up the
ethertap device from EthernetDataTest_Send. We currently have no way to
get the name or path of the ethertap device we created, so the file
watcher can occasionally see the wrong device if devmgr doesn't clean up
in time. Thus the ethernet client is talking to a different device than
the ethertap we are writing to, and the recv never makes it to the
client. Adding a fixed delay after every ethernet test gives devmgr time
to clean up the ethertap device so the next test runs without any
confusion.

These flaky tests would fail fairly quickly in qemu+kvm on a Z840,
running in a loop:
while /boot/test/sys/ethernet-test; do true; done

After this change, the tests run in a loop without failing.

NET-225 #done

Change-Id: Idf1919f7ea000ab2b52a53e1ff5c69fc1526d35f