[ledger] Fix flakiness on PageUploadTest.DoNotUploadSyncedCommitsOnRetry
The test inject whether there is unsynced commit available or not, and
check that we do not try to upload a commit that was received from the
cloud.

The issue was that uploading a commit to the cloud is asynchronous, and
that injection could happen while an upload was in progress.

The CL fix the issue by enforcing the injection between 2 uploads tries.

LE-333 #done

Change-Id: I46be25535cebdf91330b836fac43a9126f75f128