[lib][hypervisor] Calculate cpu_of correctly.
VPID always starts at 1, as both x86 and arm64 require it not to be 0
for the guest. However, it's convenient to have the first VPID map to
the first CPU, rather than be offset by 1.

For example, on arm64, this means we do not need to set VMPIDR_EL2 to
identify the primary CPU.

ZX-1347 #comment

Change-Id: I819eadd821628c1a93411da05129c8c2a7306dfd