[pci] Don't disable paradise serial during mexec
The short term workaround where we disable all pci devices
during mexec was also turning off the serial port on paradise.
Since this device is only configured by the bios it resulted
in the console being disabled entirely. Until we have the proper
hooks around disable/suspend, we'll specifically skip disabling
that device.

Also, stop disabling MMIO/PIO access to the device. With the device
and bus mastering disabled that should be enough.


Change-Id: Ie3769479f85b76e9347fbe638fc3dee921692237