[wlan][mlme] Adds FrameHandler [3]
The FrameHandler is an interface with default
implementations to handle all by the system
supported frame types, such as Data, and
Management ones.

*Dropping Frames*:
In addition, the FrameHandler provides a simple
way to drop incoming frames which should result in two
benefits, (1) a centralized way of dropping frames rather
than having the logic spread out and duplicated in various
methods, and (2) to make it easy for frame handlers, such as
Station & Scanner to take responsibility of dropping frames
it is not intersted in, rather than depending on its
parent/owner to drop frames for it, as it was the case in the
previous architecture, which required the parent to know
about the entire state and implementation of the child to
evaluate and drop frames for it, rendering the child
essentially useless without its parent. Now, these pieces are
decoupled and can easily be interchanged e.g., with mocked
implementations for testing.

*MLME Message*
Another bigger change can be found when looking at MLME
Messages. A StructPtr is a unique pointer which previously was
passed down from Device to Mlme, and Mlme to Station, etc.
It was not possible to let more than one component process
MLME messages because ownership had been passed already to
another component. This could be solved by passing raw  pointers
instead. However, this would cause another problem. For example
the Scanner holds onto the the SCAN request object. If this
would be a pointer, we could potentially run into
use-after-free or similiar scenarios. The new implementation
provides two benefits, (1) multiple components can process
a MLME message, and (2) a component can hold onto a message
without worrying that other components cannot process the
message. This is achieved by using a shared pointer.

Change-Id: I817729af2ffdb6c4e7bce05c697c11c0f0c90392