[build] Support user-only and kernel-only builds
The new user-arm64 and user-x86-64 projects can be used to build
the sysroot and/or user bootfs without building any kernel (or
choosing a target platform).

e.g. `make PROJECT=user-x86-64 user-only` will build the sysroot,
the bootfs.manifest file and everything in it (but not the user
`bootdata.bin` image itself, since Fuchsia builds don't use it).

e.g. `make PROJECT=zircon-qemu-arm64 kernel-only` will build the
kernel images and a `kernel-bootdata.bin` file that should be
combined with a bootfs image to go with this kernel.

To keep the build process uniform, `kernel-bootdata.bin` is
generated even when it's empty (which is currently always the
case for x86), hence `mkbootfs` acquires the `--empty` switch.

Change-Id: I1090b299e2a7f30b4e6afb53e785faf480dc9ea7