[sketchy] Implement a buffer pool for multi-buffering.
Multi-buffering is now handled within BufferPool. It monitors release
fences from scenic, and requires the client to return the buffer
manually.

Buffer is no longer auto-increasing. MeshBuffer becomes the client of
BufferPool, and manages the resizing of Buffers.

As demonstration, a wavy stroke will change its path every second. When
it's at top, it's shorter. As POC, BufferPool is verified to reuse the
short/long buffers interleavingly.

Change-Id: Ief879d1828912d32efef7c8ee452ab2eccd60259