[ledger] Split ledger package into pieces
Previously, the ledger package included the ledger, the Firebase cloud
provider, as well as the ledger_tool. After this CL, these are all
separate packages, which will make it easier to sandbox each of them.

Also, separate the bup_unittests into a separate package for bup down
that the code is in its own directory.

Change-Id: Ie0e1aded5318970b01ff9e8dfe259035318cc0da