[ledger] Use SizedVmo instead of raw VMO.
VMO will soon be page aligned. The Ledger was using the size of vmo
internally, as well as in its inbound and outbound APIs.

This CL change the Ledger to not use raw api anymore. It now uses
SizedVmo insternally, and SizedVmoTransport at the API level.

This CL also updates all ledger's clients in peridot for the new APIs.

Change-Id: Ibe182254be8d98c18f09b3ed2a20e60af3857d1e