[scripts][symbolize] Add exe.unstripped to build dirs
... and put it ahead of the top level build directory to make
sure the unstripped version will be found before the stripped one.
This does mean it will get searched twice, but relative to the entire
search time, the addition is small.
Plus once a file is found it's location is cached.
Plus this is only used as a fallback in case the file isn't found in ids.txt.

Also clarify the arg to --build-dir: It replaces the Fuchsia build directory.

Change-Id: I0d59d7ed4fc4088c58a1810b581d8162643c2286