[modular] Use common main for test components.
The main() functions of applications that implement components
used in integration tests using ComponentBase look similar enough
so that they should be factored out.

There was concern around AppDriver that it should not imply that
component implementations are *always* the only thing their
application does. We still not imply or encourage that here,
but simply acknowledge sitations where that's how it is.

MI4-344 #comment One more improvement.

Change-Id: I04e2b4322770e4cda51181977985707662b4af21