[modular] Construct termination actions during Initialize().
On initialization of every member that needs to participate in
the asynchronous termination sequence executed in Terminate(),
schedule an appropriate action, which is stored in a vector.

In Terminate(), just execute the stored actions in reverse order.

An action is a function that takes a callback, so it can execute
asynchronously if it needs to.

There are two types of termination actions:

* synchronous invocation of reset() on a unique pointer, or

* asynchronous invocation of Teardown() on an AsyncHolder
  or an AppClient.

With this, it is no longer necessary to explicitly keep
Initialize() and Terminate() in sync, and it becomes much
simpler to omit fields from the initialization sequence.

Collateral cleanup:

* Consistently call app client members *app_.

* Move initialization of user shell down so it matches the
  termination sequence.

MI4-682 #done

Change-Id: I8c07d248ecd90d9dd0ec50a5289537d7fd0366c1