[audio][mixer] Add a utils and constants file.
Add a dumping ground for small but handy utility objects and constants
used across the mixer code.  Start to migrate some of the existing
constants into here.  More still needs to be moved into here, but at
least there is a place to put them now (instead of scattered across
the project)

Change-Id: I8f19e96d3bd49b0b882b36a989292472e859401a