[hypervisor][x86] Fix idle on linux.
Disable MONITOR/MWAIT in CPUID so that linux will choose hlt wait. Also
clear any guest interruptibility blocking flags which is required to
support the linux hlt loop:

  sti; hlt

When we take a vmexit on the 'hlt' instruction the 'blocking by STI'
bit is still set in the guest interruptibility state. Since we emulate
this instruction it's safe to clear this flag.

ZX-1280 #done

Change-Id: I92a9840e5d08aedda7fe9c6559901bf377d7463c