[devmgr] Fix device_remove() for devices added with DEVICE_ADD_MUST_ISOLATE
If a driver adds a device with DEVICE_ADD_MUST_ISOLATE, the new device
is added in a new devhost.
Currently when this driver calls device_remove() on this device,
the device and its children are removed from the devcoordinator.
But the devhost containing the device is not informed, so drivers in
that devhost are not informed and are left still actively running.
This change adds an RPC from the coordinator to the devhost to tell
it to remove the device.

This showed up on the Hikey960. The "usbctl mode device" command
failed because the xhci driver did not shutdown so the dwc3 driver
failed to open an interrupt handle for the interrupt it shares with xhci.

Change-Id: I5c20cb072a2e3703ac028f146e216db12360cf05