[modular] Fix AsyncHolder destruction during Teardown().
If an async holder is deleted while the Teardown() is executing,
the timeout callback must be prevented from executing. We
accomplish this by making the shared bool that is already used
to prevent the timeout callback from executing after the regular
teardown callback, a member of the async holder. We then set it
in the destructor so it prevents any pending callback from
executing.

This also allows to log this situation. It's not a warning because
it happens usually when an outer teardown times out, for which
there is a warning already.

MI4-689 #done

Change-Id: I4ef91a8e43a2fd1556c7ebfe158f778898f617e4