[modular] Don't rely on unobservable state transitions.
A pipeline of API actions that causes state to transition
through ephemeral states to a stable end state cannot be
relied on to be observable on a connection separate from
the pipeline. Only the stable end state can be relied on
being observable.

Many of these transient module states will likely go away
anyway, as they are not really useful, and indeed not used.

This change was also tested with https://fuchsia-review.googlesource.com/#/c/garnet/+/89431/
patched in, and still passed on both qemu and NUC.

MI4-684 #done
MI4-304 #comment Another example for states not being too useful.

Change-Id: I221ef15e578f82df1a8ebeaee44a82d67d95801b