[media] Enable NxN mixing
This CL enables the system mix engine to handle multi-channel capturers,
as long as the number of channels matches the associated HW input device.
In this mode, channels are mapped directly through without
rechannelization or attenuation, except for normal capturer gains.

Because the current system mixer contains template specializations
that provide hardcoded optimized versions for the common 1-/2-channel
cases (one or two channels coming in, being mixed to one or two
channels going out), the changes in this CL consist of the addition
of a non-templatized version that passes the number of channels as a
parameter or member variable.

The CL includes NxN versions of both the PointSampler and the
LinearSampler, as well as the addition of a output formatter that
is more channel-flexible (not hardcoded by template to 1- or 2-
channel outputs).

This changes our maximum supported channel count, which is represented
by a constant in public/lib/media/fidl/media_types.fidl. Because this
value now differs between output and input, kMaxLpcmChannelCount has
been split into two separate constants (kMaxLpcmInputChannelCount and
kMaxLpcmOutputChannelCount).

Change-Id: Ic1c25982428df9bae94aa0ac8b64f16d921e0f8d