[async] Handle re-entrant cancellation
Previously, the loop would call the task epilogue while the loop was in
an inconsistent state. In some cases, the epilogue would call back into
the loop to cancel the current handler, which would fail because the
current handler had not yet been re-registered. When the stack unwinds,
the loop would re-register the now-cancelled handler.

US-381 #done
ZX-1385 #done

Change-Id: I76957aa10363ae98a66802e4ba733dbff75aa289