[resolver] Do the right thing when re-adding an existing manifest.
The wrong thing: remove knowledge of the entry, but add it to the
inverted index anyway, leading to either a) not seeing the Module in Ask
and/or b) crashing when trying to resolve that module.

The right thing: remove the entry from the index, and then re-add the
entry and don't just forget about it completely.

MI4-693 #done

Change-Id: Ied98fc3f63297f05dad6e9982133d555acbd5340