[audio][mixer] Get ready to add capturers.
Add some functionality to audio devices and audio drivers which will
be needed by capturers in order to be able to capture from ring buffer
sources.  Also tweak the constants in driver_output so that it holds
onto mixed frames for a longer period of time so that capturers which
wake up less frequently have a chance to capture from the output.

Change-Id: Ieb1937d75bf290c7cef64fbf735c67cc3c799b1a