[dev][usb][xhci] Handle stalls in the status phase of a control requests
Some devices (like the Schitt Fulla 2 audio interface) will report success
during the data phase of a control request but then issue a stall during
the status phase. Previously we were completing the request when we receive
the transfer event corresponding to the data phase. Later when we receive
the stall event for the status phase it is too late to report this back to
the client.

To fix this we now set the IOC bit on the status TRB so we now receive
a transfer event for both the setup or data phase and the status phase.
When we receive the first transfer event we record the result in the
USB request, but defer completing that request until we receive the
transfer event for the status phase. At that point we update the
USB request's result status if necessary and then complete the request.

ZX_1375 #comment in progress

Change-Id: Ib7a1c4164f0ecf09b75735293aa1985465cd2a4b