[fidl][host] Tweak the fidl host tool build layout
Many of these files are consumed by a currently disabled Garnet gn
target (it is disabled until a future toolchain roll fixes some
fuzzing support infrastructure), which essentially builds a
fuzzing-oriented version of the compiler.

The current layout is expressed as a BUILD.gn file in Zircon. When we
remove these, in favor of generated manifests, this layout will be
difficult to express. The new one puts everything but main into a
separate directory.

Change-Id: I18ea1807fa5d75a3b2620f3a54d4aac4999e0442