[ledger] Remove timeouts from PageStorageTest
This is a proof-of-concept that can probably be applied to other tests.
Although test timeout flakes are rare, they do happen, and I want to
show that it's not prohibitively difficult to completely eliminate the
possibility in some kinds of tests.

Any test that runs on a single thread shouldn't need to set a timeout to
wait for conditions. By using these techniques, we just don't have to
worry about choosing the appropriate timeout for single-threaded tests.

Changes are as follows:

- Inject a FakeTaskRunner into PageStorageImpl. FakeTaskRunner stores
  tasks in a queue and can run them when requested. This replaces the
  global MessageLoop. ppi@ asked on a previous version of this CL if we
  could add a RunUntilIdle() method to MessageLoop. That would also
  work, but FakeTaskRunner has the added advantage of being
  explicitly provided as a dependency, i.e., not a global object.

- Remove the base class, StorageTest, and re-implement the necessary
  helpers using FakeTaskRunner.

- Verify that callbacks were called using SetWhenCalled(), instead of
  asking whether MessageLoop::Run() exited before timing out.

- Change AddObjectFromLocalWrongSize (which uses SocketDataSource with
  the wrong size specified to create an error condition) to
  AddObjectFromLocalError (which uses FakeErrorDataSource to explicitly
  return an error). This avoids the use of fsl::SocketDrainer, which
  depends on the global MessageLoop.

LE-358

Change-Id: I2473ab695357792e38c4feb5b74c4ef3ff7b4f90