[blobstore] Allow host tool to add multiple files at once
Additionally, on create if any manifests/blobs are specified, truncate
the blobstore to a size large enough to fit all of them.

This change will update the blobstore create/add calls to the following:

blobstore <file> create --blob <path-to-blob> --manifest <path-to-manifest>
blobstore <file> add --blob <path-to-blob> --manifest <path-to-manifest>

--blob and --manifest args are optional for create, but at least one is
required for add. Adding a blob and a manifest are no longer two separate
commands.

Change-Id: I05781cc529c30ae48ce425ea799b204dd45f6eb2