[audio][mixer] Select the "best" input/output mode.
Attempt to select the "best" input or output mode from the list of
modes the driver gives to us as supported modes.  Note: the heuristic
for "best" implemented here is a temporary thing.  Eventually we need
a policy manager which takes into account all of the user
preferences and requirements (system wide), along with the driver
capabilities, and decides which mode an input or output should come up
in, and if/when it should change modes to something else.

Change-Id: If0600a93afec88513e0b99180edb0ee955a3a4f3