[arm64][hypervisor] Save and restore EL0 state
Add TPIDR_EL0, TPIDRRO_EL0, and SP_EL0 to SystemState, so that we save
and restore the correct guest state and host state.

This fixes the bug where TPIDR_EL0 is not correctly restored for the
host, so when we return to user-space and continue execution, we crash
as soon as TLS is used.

ZX-1347 #comment

Change-Id: If0912931e7060ce328766c16894883ef18be279b