[modular] Send non-embedded ancestors to story shell as anchor views.
When a module is started with StartModuleInStoryShell(), the story
shell is provides with a module relative to which the newly displayed
module is anchored. This view was formerly called the parent view,
because it was always the parent module. However, if the parent
module is embedded by its own parent in turn, the story shell doesn't
know about the parent and will not display the new module.

Therefore, for the child of an embedded module, we use the closest
non-embedded ancestor module as the ancestor for the story shell
display.

The mechanism to defer displaying of a module until its anchor is
known to the story shell is adjusted accordingly.

Add tests for the new behavior. This for the first time uses a
story shell as part of a test, and uncovers an interference of
story shell preloading with test points: The extra preloaded copy
of the story shell doesn't pass its test points (of course). Thus,
in tests, disable story shell preloading.

Collateral bugfixes:

* Pending views of a module is closed when the module is
  torn down.

* Tear down the preloaded copy of the story shell during
  story provider teardown.

MI4-688 #comment Open child of embedded module in story shell

Change-Id: Ic18b737f7257ea6e8b8600b9e010c65d06a919d7