[wchar_t] Fix the definition of WCHAR_MIN and add tests
This adds a definition of WCHAR_MIN that works in C and C++, and under
GCC and Clang. The two compilers provide different predefined macros
for libc's need to provide this value.

This patch also duplicates a lot of the existing stdint tests to make
sure things work in both C and C++, and that they work when only
<wchar.h> is included.

Change-Id: I90bc0aed3dd921015adbca2d8ec6cd43ce3c737a