[cloud_provider_firestore] teach FactoryImpl to fetch the user id
This patch implements retrieving the Firebase user ID in the cloud
provider factory, so that each instance of the cloud provider is
configured with the ID of its user. This is a prerequisite to be able to
construct correct Firestore paths in order to make requests.

LE-368 #comment [cloud_provider_firestore] teach FactoryImpl to fetch the user id

Change-Id: I9ae77fac517f3bcabebeb3c53a8e34ace69db7c7