[benchmarks] Add new entry point for running the benchmarks as fast tests
Add a new executable, zircon_benchmarks_test.  The initial purpose of
this is to run the benchmarks on the bots, via runtests, to ensure
that they keep working.  Initially only the tests in round_trips.cc
are run.  I will add the others in a later change.

For zircon_benchmarks_test, we add some new code for registering and
running tests that bypasses the gbenchmark library.  We do this for
these reasons:

 * By default, gbenchmark is very slow, which isn't ideal for
   runtests.  In contrast, zircon_benchmarks_test runs only a small
   number of iterations of each test, so it's fast.

 * gbenchmark has a problem where it ignores errors: If a test fails
   with SkipWithError(), gbenchmark will return a exit status of 0,
   which isn't what we want for runtests.  zircon_benchmarks_test will
   avoid this problem.

 * Later this will become more of a replacement for gbenchmark by
   running the tests more times and logging the raw timing data for
   analysis (e.g. via the Fuchsia tracing framework).

Change-Id: I3d8fd7f67577f0ac6173ed129d8335903153d9fe