[wlan][ap] Adds BeaconSender
The BeaconSender is responsible for sending Beacon frames reliably for a
given configuration. The Sender currently only supports a single BSS
which might need to evolve in future development. For now this should be
sufficient though.
Beacons are sent from a separate thread. The thread's priority might
need to be increased if the beacons are not send reliably enough. That
however needs to be evaluated first.

NET-251

Change-Id: I96df02e93c72137ffea050bd8647ad70783f7a8c