[modular] Transitive Embedding controllers
When a module embeds another module, and that module starts
more modules in the story shell, the first module can receive
notifications of it, which allows it to focus and defocus those
transitive modules.

This is done with a new method ModuleContext.EmbedModule(), but
this should really become the default for embedding.

EmbedModule() takes an EmbedModuleWatcher, on which
OnStartModuleInShell() is called for every module started transitively
in the story shell. The callback receives an EmbedModuleController
connection to the module controller which, unlike the full
ModuleController, only allows to Focus() and Defocus() the module,
but not Watch() or Stop() it.

Add a test for this machinery.

Collateral fix: {Embed,}ModuleController.Focus() now sends the
right anchor view id to story shell in the case that the parent
module is embedded.

MI4-688 #done

Change-Id: Ifcb9b23a3f49b3eedce12b690cf8a199dfbc6550