[wlanstack] Sandbox wlanstack
The wlanstack has access to /dev/class/ethernet. We also give it access
to /system/data/wlanstack because some builds place a config file in
that directory. We'll eventually want to find a way to provision the
wlan configuration that doesn't involve the system package. Once we do,
we can remove access to the system package.

Change-Id: Ia5c506638a3f4d975cd700994ffa968de49e23f6