[media] Add a config file to control whether the mixer owns audio input
The way inputs are handled in the media subsystem is changing, but it's
difficult to update all users of this system at the same time. Briefly,
the old path involved the media service owning the audio input devices
and providing input streams to clients. The new path has the mixer own
the audio input devices and provide input streams to clients through a
slightly different interface.

We currently have different clients using both kinds of interfaces that
we want to be able to produce working builds for until we can convert
everything to the new path. To accomodate that, this adds a package
config that drops an (empty) file into /system/data/media that the
mixer can check for to decide whether to enable the new path. We can
set this in configurations using the new path and leave the old path
alone.

Change-Id: I0de2e5468f8786bd8933938b46bf20b8d02665ed