[machina][gpu] Use skia for resource management.
Restructure the gpu to utilize skia types for drawing and resource
management.

  * Each GpuScanout will have an SkSurface object that will be the
    render target. To support the zircon framebuffer we simply create a
    surface that wraps the linear framebuffer.
  * Each GpuResource will have an SkBitmap that holds the raw pixel
    data (updated via transfer 2D commands).
  * Pixel data is drawn to the SkSurface from the SkBitmap whenever the
    guest sends resource flush command.

The previous implementation took a shortcut by providing the
framebuffer directly to the resource object. This worked fine for
simple use cases (ex: the linux & zirxon virtcons) but wouldn't support
more complex use cases that X uses.

This will also make is fairly straitforward to port the GPU to target
a mozart view which can also be wrapped in a skia surface.

Change-Id: I80f78fc4dcb22121bfae9b3e8c12df330cb2d7f0