[benchmarks] Add test for the round trip time of FIDL IPC
This uses synchronous IPC on the client side.  That's mostly because
that fits best into the current perf testing framework that is used in
round_trips.cc where each iteration of the perf test is done as a
separate Run() function call.

In future we can add a performance test of asynchronous IPC.  For that
we can either use Quit() to exit the MessageLoop on each iteration, or
we can change the testing framework so that we can accumulate timing
results without leaving the MessageLoop.

Change-Id: I2e263e7b00d6c1222a45811f23dfbe7f099d9f40