[wlan] Generalizes BssMap into MacAddrMap
Changes to BssMap before extracting it into MacAddrMap:
(1) BssMap holds onto the Bss via RefPtr to prevent ownership issues
(2) Pruning is moved out of the Map and into the Scanner. It's was too
tightly coupled to the Scanner's use-case
(3) Made the limit optional since only the Scanner Map must be limited
(4) Removed direct access to the underlying map
(5) Removed Upsert which was too coupled to pruning.
(5) Removed HasKey(). Lookup() returns nullptr if it wasn't found.

Change-Id: I72f729421a959b2e117df3cbb61cd7866ed652e2