[audio] Update audio driver protocol
Update audio driver stream protocol.  Significant changes now include...

1) Drivers may now report an external delay (expressed in nanoseconds)
2) DMA start times have been switched from ticks to clock monotonic.
3) Effective ring buffer size is now carried out of band (preparing to
   the up-and-coming change to zx_vmo_get_size)

Also, default initialize all protocol responses to all zeros.  Generally, the
idea here is to make sure that all our response structures have a know
initialization in them, so that if the members of the structures are updated in
the future, we won't introduce any non-deterministic bugs.

Change-Id: Ic281c889d3dbf352c59c919ebf92d98baf28e7f8