[dev][usb-bt-hci] Break wait on channel change
A race condition existed a channel was opened while another channel was
being waited on - the new channel would not be monitored until the first
channel had activity.

Added an event to break the wait and update the wait item list when
the event is signaled.

Minor cleanup:
 - DRY *_channel_cleanup_locked -> channel_cleanup_locked
 - DRY hci_ioctl channel opening code
 - Correct comment in bt-hci.h snoop flags

NET-305 #done

Change-Id: I9421ac0e86e692c22f5647ee5013cc6174fb46d5