Produce a minimal BOOTFS manifest for Zedboot images
This limits the files included in the image to those that are
necessary for running Zedboot.  (Actually, it includes all of
Zircon's drivers, though only some of them are actually useful
for Zedboot.)

This also provides a handy "zedboot" GN group so that just
`ninja -C ... zedboot` will create all the Zedboot images.

TO-611

Change-Id: I7dcccd5b49532f3609d54f07247847203a588637