[build] Simplify target ("project") names
The only projects supported now are Zircon, so having "zircon-"
at the front of every name is just more typing.  There is only
one architecture associated with each platform name, and for
x86-64 there is only one platform.  The old -> new mappings are:

	zircon-pc-x86-64	->	x86
	zircon-qemu-arm64	-> 	arm64
	zircon-gauss-arm64	-> 	gauss
	zircon-hikey960-arm64	-> 	hikey960
	zircon-odroidc2-arm64	-> 	odroidc2

The old project names continue to work as aliases during
a transition period but they are deprecated.

Since all projects are Zircon (nee Magenta), the -m switch to
buildall is removed and the build-all-zircon script is deprecated.

Change-Id: Icf88e93b460df053be251c88d9dffd8c6da82518