[time_service] Use ICU data for timezones.
The timezone is no longer set as a raw integer offset, but an ICU
timezone ID. tz-util is updated for this change as well.

The FIDL interface has updated so that getting the local time offset
requires passing a Unix time in milliseconds, so that a user
may query for times in the past or future in addition to the present.
This includes calculation of DST.

Change-Id: I2b6898bd72c46a575bd1555e6ba7a9559d3b596f