[memfs][devmgr] Refactor MemFS into a library
Relocates the MemFS implementation from devmgr and splits it into
different files, depending on which Vnode type is being implemented.

Additionally, removes the unused "ioctl_vfs_mount_bootfs_vmo",
which tightly coupled the MemFS implementation with the devmgr.

Additional changes for ulib/memfs would be useful (splitting
the header, adding tests, adding documentation, etc), but
those changes are not included in this patch so the relocation
can remain simplified.

Change-Id: I66413215a4000565865c6394a09180922d82e16a