[appmgr] Add system-temp and improve vulkan sandbox features
Using vulkan also requires access to /dev/class/gpu. This CL add that
access to the vulkan feature.

Some applications use the global system temp directory. We should
eventually remove the global system temp directory, but for now this CL
introduces a sandbox feature to gate access to the global system temp
directory so that we can whitelist clients during the removal process.

Change-Id: Ief382db130433d5fe7f70b499fb902949420d1bc