[context] Hang any Entities we see in a Link off its LINK node.
This change stores any Entities we see in a given Link off its
corresponding node in the context engine.

Since we still have JSON-style Entities around (where there's an "@type"
property in the object) as well as no Bundles yet (where the Link value
is an object with multiple members, each of which can be an Entity
itself), we have to handle a few different cases. This will be much
simplified when we have both Bundles and Entities everywhere.

MI4-487 #comment [context] Hang any Entities we see in a Link off its LINK node.

Change-Id: Ib29ce36481305b916b7fc65c31e5b61a32982a85