[libc][stubs] Correct the stub definition of freeaddrinfo
When the real definition of freeaddrinfo (and other associated
functions) was moved to fdio, the libc source file definining
freeaddrinfo was left in the tree, even though there was no build rule
for it.

Additionally, the stub_freeaddrinfo function was not marked static.

This CL fixes both of these glitches.

Change-Id: If7a4f10f9faa6c3223a8a9c7c5fc6a9d9cfb12de