[media] Sandbox audio_servier and media_service
These apps need access to audio devices to work properly.

Also, audio_server needs access to a config file in the system image in
order to determine whether it should take control of the audio input
drivers. Once that config file is no longer needed, we should remove
access to the system package.

Change-Id: I7c892eae195cd1fe29fb5759d2ec10c149f0c123