[kernel][counters] Use a pointer for k_counter_desc.name
Indirection means the fixed-sized struct can stay small while the
length of counter names is unbounded and short names don't waste
space for padding.  The binary search depends on the order of the
structs matching the sort order of their names, but that ordering is
controlled by the linker sorting of the section names and is
orthogonal to where the name strings themselves appear at runtime.

With the current small set of counters this makes the kernel image
slightly smaller.  If the majority of counter names stay shorter
than 48 bytes, it will always be size win.

Change-Id: Ibdb76922677be48b7abdf0ed11dbca4f59925ffe