[audio][capture] Relax restrictions on VMO size
Do not demand that the size of the VMO used as the shared capture
buffer be a perfect multiple of audio frames.  Also, do not demand
that the number of frames produced per packet while operating in
'async' mode divide the overall shared buffer size.  Instead, simply
demand that a user be able to fit at least two async packets in the
shared buffer.

Change-Id: I59bf65642d9c39bdbd61155f28224ec5196e36f1