[zircon][process] don't kill process on handle
.. on handle close

This is the second step, started here
https://fuchsia-review.googlesource.com/c/zircon/+/95340

With the in-depth info in ZX-820 and in ZX-1106

So up to this point closing the last handle to
a process killed the process. With this change this is
no longer true. Processes still can be killed via the
job and the job semantics have not changed and
will not change; closing the last handle to a job
does tear down the job tree.

Change-Id: Ieb86899a95fc4dcd715b4c3c954c61db6f7ec6d6